Fixed an issue where redux was not sending a mapStateToProps call whenever an action was being called from componentWillMount, instead it rolled up all of the mapStateToProps calls into one.  Instead, we moved the same call to componentDidMount which results in the mapStateToProps call being called after the action and the selected item being set.
We also had to put in a check whenever we're retrieving the panel to
look for whether the selected item has been set.  The page will attempt
to render after the component mounts and our action is dispatched after
this.  This results in a second render and loading panel being shown
while the action is worked, but it's better than the alternative of the
panel not loading at all.